+++ src/duk_bi_string.c
@@ -531,8 +531,8 @@ duk_ret_t duk_bi_string_prototype_replace(duk_context *ctx) {
 
 				duk_get_prop_stridx(ctx, 0, DUK_STRIDX_LAST_INDEX);
 				last_index = (duk_uint32_t) duk_get_uint(ctx, -1);
-				DUK_DDD(DUK_DDDPRINT("empty match, bump lastIndex: %d -> %d",
-				                     (int) last_index, (int) (last_index + 1)));
+				DUK_DDD(DUK_DDDPRINT("empty match, bump lastIndex: %ld -> %ld",
+				                     (long) last_index, (long) (last_index + 1)));
 				duk_pop(ctx);
 				duk_push_int(ctx, last_index + 1);
 				duk_put_prop_stridx(ctx, 0, DUK_STRIDX_LAST_INDEX);
@@ -988,10 +988,10 @@ duk_ret_t duk_bi_string_prototype_split(duk_context *ctx) {
 		 * stack[4] = regexp res_obj (if is_regexp)
 		 */
 
-		DUK_DDD(DUK_DDDPRINT("split; match_start b=%d,c=%d, match_end b=%d,c=%d, prev_end b=%d,c=%d",
-		                     (int) match_start_boff, (int) match_start_coff,
-		                     (int) match_end_boff, (int) match_end_coff,
-		                     (int) prev_match_end_boff, (int) prev_match_end_coff));
+		DUK_DDD(DUK_DDDPRINT("split; match_start b=%ld,c=%ld, match_end b=%ld,c=%ld, prev_end b=%ld,c=%ld",
+		                     (long) match_start_boff, (long) match_start_coff,
+		                     (long) match_end_boff, (long) match_end_coff,
+		                     (long) prev_match_end_boff, (long) prev_match_end_coff));
 
 		duk_push_lstring(ctx,
 		                 (const char *) (DUK_HSTRING_GET_DATA(h_input) + prev_match_end_boff),
@@ -1032,8 +1032,8 @@ duk_ret_t duk_bi_string_prototype_split(duk_context *ctx) {
 
 	/* Combined step 11 (empty string special case) and 14-15. */
 
-	DUK_DDD(DUK_DDDPRINT("split trailer; prev_end b=%d,c=%d",
-	                     (int) prev_match_end_boff, (int) prev_match_end_coff));
+	DUK_DDD(DUK_DDDPRINT("split trailer; prev_end b=%ld,c=%ld",
+	                     (long) prev_match_end_boff, (long) prev_match_end_coff));
 
 	if (DUK_HSTRING_GET_CHARLEN(h_input) > 0 || !matched) {
 		/* Add trailer if:
